pyproject.toml
[tool.poetry]
name = "care-classifier"
version = "0.1.0"
description = ""
authors = ["supersweet-dev <pentagr4mi@gmail.com>"]
readme = "README.md"
packages = [{include = "care_classifier"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
strawberry-graphql = {extras = ["channels"], version = "^0.142.3"}
strawberry-graphql-django = "^0.5.1"
Django = "^4.1.1"
uWSGI = "^2.0.20"
django-uwsgi = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
